Communication tests
This commit includes a variety of ways of communicating with an application running on a cell.  It includes both sync and async as well as threads, channels, and select within the code running on the cell.

It also includes code from a couple of failed experiments that pass file handles of pipes from the app_agent process to the client process.  It worked for a sync server sending to a client running natively, but failed for a client running in a Docker container; the container could not see the Unix stream used to pass the file handles.  It also failed for an async server sending to a sync client; the async file handles appear not to be valid sync file handles.